requirements.txt
flake8
importlib_resources
pynini==2.1.6
pytest
pre-commit==3.5.0
opencc-python-reimplemented
